createSurface bug fix
  This fixes a number of issues:
    (i) createSurface [in surIndex] builds over an existing surface
        In this case the surface ptr which is used elsewhere is invalid

    (ii) update of the surface is needed to update the object/rules
    that use the surface

 This is TEMPORARY corrected by addint updateSurface(int,string)
 to Simulation. This deals with the object update.

 HOWEVER: This should be fixed by (a) moving surIndex / surfRegister
 into one class. Providing only smart pointers to the surfaces and then
 when a surface is updated the surface will be automatically updated
 everwhere.

+++ Model/delft/beamSlot.cxx
@@ -228,7 +228,8 @@ beamSlot::createSurfaces(const attachSystem::FixedComp& FC)
   ModelSupport::buildPlane(SMap,surfIndex+14,
   			   Origin+X*(xSize/2.0-endThick),X);
 
-  const double gap=(zSize-(NChannels+1.0)*divideThick)/NChannels;
+  const double gap=(zSize-static_cast<double>(NChannels+1)*divideThick)/
+    static_cast<double>(NChannels);
   double zPoint(-zSize/2.0);
   int surfOffset(surfIndex+10);
   for(size_t i=0;i<NChannels;i++)
